backend/pyproject.toml
[project]
name = "tfg-backend"
version = "0.1.0"
description = "Backend FastAPI - CMDB, inventario y auditoría"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",  # Para SQLite async
    "pydantic>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uvicorn]
app = "app.main:app"
host = "0.0.0.0"
port = 8000
